update sentence cloud test
